Add peer.py script which creates a peer
This script is to be called by the parent script as a subprocess.
This is to make sure that each peer will run on a different
process than a thread for reliability and performance reasons.
Pass a URL to our signalling server to peer.py argument, it will
register on the server and return its UUID on stdout.
Send a UUID on its stdin after this to establish a connection with
that peer.
On the different callbacks, a text message will be sent on stdout.